[FIX]검색기능을 주소에 기반해서 실행
